+++ setup.cfg
@@ -29,8 +29,6 @@ install_requires =
 
 [options.extras_require]
 test =
-    localstack==4.0.3
-    localstack-utils==1.0.1
     pytest==8.3.4
     pytest-cov==6.0.0
     uuid==1.30
